Restructure multithread effect application
